start adding pragma
