+++ libs/core/src/lib.rs
@@ -1,6 +1,4 @@
-pub mod position;
 pub mod severity;
 pub mod span;
 pub mod store;
 pub mod symbol;
-pub mod token;
